feat: делает частичное дробление страниц на компоненты
