+++ src/const.ts
@@ -1,3 +1,12 @@
 export const DataMain = {
   RentalOffers: 312,
 } as const;
+
+export const SixCities = {
+  Paris: 'Paris',
+  Cologne: 'Cologne',
+  Brussels: 'Brussels',
+  Amsterdam: 'Amsterdam',
+  Hamburg: 'Hamburg',
+  Dusseldorf: 'Dusseldorf'
+} as const;
