+++ src/pages/favorites/favorites-empty.tsx
