+++ src/pages/login/login.tsx
@@ -1,17 +1,9 @@
+import Header from '../../components/layout/header/header';
+
 export default function Offer(): JSX.Element {
   return (
     <div className="page page--gray page--login">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <a className="header__logo-link" href="main.html">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width={81} height={41} />
-              </a>
-            </div>
-          </div>
-        </div>
-      </header>
+      <Header isNavActive={false} />
       <main className="page__main page__main--login">
         <div className="page__login-container container">
           <section className="login">
