+++ src/pages/main/main.tsx
@@ -1,0 +1,63 @@
+import Header from '../../components/layout/header/header';
+import PlaceCard from '../../components/place-card/place-card';
+import { SixCities } from '../../const';
+
+type MainProps = {
+  dataMain: {
+    RentalOffers: number;
+  };
+}
+
+export default function Main({ dataMain }: MainProps): JSX.Element {
+  return (
+    <div className="page page--gray page--main">
+      <Header isLogin isNavActive />
+
+      <main className="page__main page__main--index">
+        <h1 className="visually-hidden">Cities</h1>
+        <div className="tabs">
+          <section className="locations container">
+            <ul className="locations__list tabs__list">
+              {Object.values(SixCities).map((cities) => (
+                <li className="locations__item" key={cities}>
+                  <a className="locations__item-link tabs__item" href="#">
+                    <span>{cities}</span>
+                  </a>
+                </li>
+              ))}
+            </ul>
+          </section>
+        </div>
+        <div className="cities">
+          <div className="cities__places-container container">
+            <section className="cities__places places">
+              <h2 className="visually-hidden">Places</h2>
+              <b className="places__found">{dataMain.RentalOffers} places to stay in Amsterdam</b>
+              <form className="places__sorting" action="#" method="get">
+                <span className="places__sorting-caption">Sort by</span>
+                <span className="places__sorting-type" tabIndex={0}>
+                  Popular
+                  <svg className="places__sorting-arrow" width={7} height={4}>
+                    <use xlinkHref="#icon-arrow-select" />
+                  </svg>
+                </span>
+                <ul className="places__options places__options--custom places__options--opene-d">
+                  <li className="places__option places__option--active" tabIndex={0}>Popular</li>
+                  <li className="places__option" tabIndex={0}>Price: low to high</li>
+                  <li className="places__option" tabIndex={0}>Price: high to low</li>
+                  <li className="places__option" tabIndex={0}>Top rated first</li>
+                </ul>
+              </form>
+              <div className="cities__places-list places__list tabs__content">
+                <PlaceCard isCitiesCard />
+              </div>
+            </section>
+            <div className="cities__right-section">
+              <section className="cities__map map"></section>
+            </div>
+          </div>
+        </div>
+      </main >
+    </div >
+  );
+}
